Adjusting main with the two py.scripts
