+++ Main.py
@@ -9,6 +9,7 @@ from time import time, sleep
 import numpy as np
 import logging
 from tabulate import tabulate
+import pymysql
 
 #--------------------------------- Initialisation of charging stations, battery and grid -----------------------------------------------------------------#
 
@@ -99,6 +100,8 @@ for i in range(all_inputs.shape[0]):
     total_charging_cost = total_charging_cost + charging_cost
     total_charging_profit = total_charging_profit + charging_profit   
    
+    ########################################################################################################################
+    # Sefa's variable calculations ################
     pv_elec_slice_openwb = (openwb.pv_elec / openwb.charging_power) * 100 if openwb.charging_power !=0 else 0
     pv_elec_slice_webasto = (webasto.pv_elec / webasto.charging_power) if webasto.charging_power !=0 else 0
     residual_time_openwb = openwb.charge_duration - counter if openwb.charge_duration != 0 else 0
@@ -118,47 +121,87 @@ for i in range(all_inputs.shape[0]):
     P_house_rounded = round(P_house,2)
     used_battery = round(hbattery.used_power,2)
     battery_soc = round(hbattery.soc / hbattery.soc_max,2) * 100
+
+    ############################################# Addition for User Interface #############################
+
+    #database connection
+    connection = pymysql.connect(host="localhost", user="root", passwd="", database="test")
+    cursor = connection.cursor()
+    # Query for creating table
+    HEMSTableSql = """CREATE TABLE HEMS_values(
+    ID INT(20) PRIMARY KEY AUTO_INCREMENT,
+    VARIABLENAME varchar(50),
+    VARIABLEVALUE numeric(9,2),
+    VARIABLEUNIT varchar(50)
+    )"""
+
+    cursor.execute(HEMSTableSql)
+    connection.close()
     
-    l = [["Auto - Angeschlossen", openwb.connection_state, webasto.connection_state], 
-         ["Ladeprioritaet", openwb.charge_priority, webasto.charge_priority], 
-         ["Ladeleistung [kW]", openwb.charging_power, webasto.charging_power], 
-         ["PV - Strom Anteil [%]", pv_elec_slice_openwb, pv_elec_slice_webasto ],
-         ["Rest - Zeit [m]", residual_time_openwb, residual_time_webasto],
-         ["Durchschnittpreis Strom [c./ kWh]", c_elec_frame_openwb, c_elec_frame_webasto],
-         ["Gefragte Energiemenge [kWh]", openwb.e_demand, webasto.e_demand],
-         ["Gefragte maximale Energiemenge [kWh]", openwb.e_max_demand, webasto.e_max_demand],
-         ["Ladezustand (Maximale Energie) [%]", charging_level_openwb, charging_level_webasto]]
-   
-    table_charge_stations = tabulate(l, headers =['Ladestationen', 'OpenWB', 'Webasto'], tablefmt='orgtbl')
-       
-    m = [["Aktueller Strompreis [c./ kWh]", c_elec_rounded],
-         ["Aktuelle Stromnetzverwendung [kW]", grid_usage],
-         ["Aktuelle Stromnetzverwendung für Laden [kW]", grid_usage_for_charging],
-         ["Aktuelle Ladekosten [€] ", charging_cost_rounded],
-         ["Aktueller Gewinn durch HEMS [€]", charging_profit_rounded],
-         ["Gesamte Ladekosten [€]", total_charging_cost_rounded],
-         ["Gesamter Gewinn durch HEMS [€]", total_charging_profit_rounded]]
-         
-    table_balance_sheet = tabulate(m, headers =['Bilanz', 'Wert'], tablefmt='orgtbl')
- 
-    n = [["PV Energie Erzeugung [kW]", P_pv_rounded],
-         ["Haushalt Stromverbrauch [kW]", P_house_rounded],
-         ["Heimspeicher Leistung [kW]", used_battery],
-         ["Heimspeicher Ladezustand [%]", round(hbattery.soc / hbattery.soc_max,2) * 100]]
-         
-    table_house = tabulate(n, headers =['Haus', 'Wert'], tablefmt='orgtbl')    
     
-    print("---------------------------------------------------------------------------------------------------------------------------")
-    print("---------------------------------------------------------------------------------------------------------------------------")
-    print(table_charge_stations)
-    print("---------------------------------------------------------------------------------------------------------------------------")
-    print("---------------------------------------------------------------------------------------------------------------------------")
-    print(table_balance_sheet)
-    print("---------------------------------------------------------------------------------------------------------------------------")
-    print("---------------------------------------------------------------------------------------------------------------------------")
-    print(table_house)
-    print("---------------------------------------------------------------------------------------------------------------------------")
-    print("---------------------------------------------------------------------------------------------------------------------------")
+    hems_verbindung1 = 1 # Intentioanlly set hard-coded, will be changed
+    hems_verbindung2 = 1 # Intentionally set hard-coded, will be changed 
+    openwb_connection_state = openwb.connection_state
+    webasto_connection_state = webasto.connection_state
+    openwb_charge_priority = openwb.charge_priority
+    webasto_charge_priority = webasto.charge_priority
+    openwb_charging_power = openwb.charging_power
+    webasto_charging_power = webasto.charging_power
+    pv_energy_share_webasto = pv_elec_slice_webasto
+    pv_energy_share_openwb = pv_elec_slice_openwb
+    remaining_time_webasto = residual_time_webasto
+    remaining_time_openwb = residual_time_openwb
+    demand_energy_openwb = openwb.e_demand
+    demand_energy_webasto = webasto.e_demand
+    charging_state_openwb = charging_level_openwb
+    charging_state_webasto = charging_level_webasto
+    Strompreis_aktuell = c_elec_rounded
+    Stromnetzverwendung_aktuell = grid_usage
+    StromnetzverwendungLaden_aktuell = grid_usage_for_charging
+    Ladekosten_aktuell = charging_cost_rounded
+    Gewinn_HEMS = charging_profit_rounded
+    Gewinn_HEMS_gesamt = total_charging_profit_rounded
+    ErzeugungPV_Energie = P_pv_rounded
+    Haushalt_Stromverbrauch = P_house_rounded
+    Heimspeicher_Leistung = used_battery
+    Heimspeicher_Ladezustand = battery_soc
+
+    #database connection
+    connection = pymysql.connect(host="localhost", user="root", passwd="", database="test")
+    cursor = connection.cursor()
+
+    #executing the queries for inserting values
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('1', 'hems_verbindung1', %s, '-' ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(hems_verbindung1, hems_verbindung1))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('2','hems_verbindung2', %s, '-')ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(hems_verbindung2, hems_verbindung2))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('3','openwb_connection_state', %s, '-'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(openwb_connection_state, openwb_connection_state))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('4','webasto_connection_state', %s, '-'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(webasto_connection_state, webasto_connection_state))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('5','openwb_charge_priority', %s, '-'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(openwb_charge_priority, openwb_charge_priority))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('6','webasto_charge_priority', %s, '-'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(webasto_charge_priority,webasto_charge_priority ))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('7','openwb_charging_power', %s, 'kW'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(openwb_charging_power,openwb_charging_power ))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('8','webasto_charging_power', %s, 'kW'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(webasto_charging_power,webasto_charging_power ))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('9','pv_energy_share_webasto', %s, '%'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(pv_energy_share_webasto, pv_energy_share_webasto))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('10','pv_energy_share_openwb', %s, '%'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(pv_energy_share_openwb, pv_energy_share_openwb))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('11','remaining_time_webasto', %s, 'min'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(remaining_time_webasto, remaining_time_webasto))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('12','remaining_time_openwb', %s, 'min'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(remaining_time_openwb, remaining_time_openwb))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('13','demand_energy_openwb', %s, 'kW'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(demand_energy_openwb, demand_energy_openwb))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('14','demand_energy_webasto', %s, 'kW'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(demand_energy_webasto, demand_energy_webasto))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('15','charging_state_openwb', %s, '%'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(charging_state_openwb, charging_state_openwb))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('16','charging_state_webasto', %s, '%'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(charging_state_webasto, charging_state_webasto))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('17','Strompreis_aktuell', %s, 'EUR'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(Strompreis_aktuell, Strompreis_aktuell))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('18','Stromnetzverwendung_aktuell', %s, 'kW'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(Stromnetzverwendung_aktuell, Stromnetzverwendung_aktuell))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('19','StromnetzverwendungLaden_aktuell', %s, 'kW'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(StromnetzverwendungLaden_aktuell, StromnetzverwendungLaden_aktuell))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('20','Ladekosten_aktuell', %s, 'EUR'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(Ladekosten_aktuell, Ladekosten_aktuell))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('21','Gewinn_HEMS', %s, 'EUR'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(Gewinn_HEMS, Gewinn_HEMS))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('22','Gewinn_HEMS_gesamt', %s, 'EUR'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(Gewinn_HEMS_gesamt, Gewinn_HEMS_gesamt))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('23','ErzeugungPV_Energie', %s, 'kW'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(ErzeugungPV_Energie, ErzeugungPV_Energie))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('24','Haushalt_Stromverbrauch', %s, 'kW'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(Haushalt_Stromverbrauch, Haushalt_Stromverbrauch))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('25','Heimspeicher_Leistung', %s, 'kW'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(Heimspeicher_Leistung, Heimspeicher_Leistung))
+    cursor.execute("""INSERT INTO HEMS_values(id, VARIABLENAME, VARIABLEVALUE, VARIABLEUNIT) VALUES('26','Heimspeicher_Ladezustand', %s, '%'  ) ON DUPLICATE KEY UPDATE VARIABLEVALUE=%s""",(Heimspeicher_Ladezustand, Heimspeicher_Ladezustand))
+
+    #commiting the connection then closing it.
+    connection.commit()
+    connection.close()
+
     
     counter = counter + 1
     sleep(60 -time() % 60)
